+++ src/RdhUtils.cxx
@@ -104,7 +104,7 @@ int RdhBlockHandle::printSummary() {
 
   int rdhcount = 0;
   RdhHeaderPrinted = false; // re-print header for each page
-    
+
   for (;;) {
 
     // check enough space for RDH
@@ -116,19 +116,19 @@ int RdhBlockHandle::printSummary() {
 
     rdhcount++;
     int offset = ptr - (uint8_t *)blockPtr;
-    
+
     const bool dumpRaw = false;
     if (dumpRaw) {
       printf("*** RDH #%d @ 0x%04X = %d\n", rdhcount, offset, offset);
 
       // print raw bytes
       // printf("Raw bytes dump (32-bit words):\n");
-      for (int i = 0; i < sizeof(o2::Header::RAWDataHeader) / sizeof(int32_t);
-           i++) {
-	if (i % 8 == 0) {
+      for (unsigned int i = 0;
+           i < sizeof(o2::Header::RAWDataHeader) / sizeof(int32_t); i++) {
+        if (i % 8 == 0) {
           printf("\n");
-	}
-	printf("%08X ", (int)(((uint32_t *)ptr)[i]));
+        }
+        printf("%08X ", (int)(((uint32_t *)ptr)[i]));
       }
       printf("\n\n");
     }
@@ -142,7 +142,7 @@ int RdhBlockHandle::printSummary() {
     }
 
     // check enough space to go to next offset
-    if (bytesLeft < next) {
+    if ((int)bytesLeft < next) {
       printf("page too small, %zu bytes left! need at least %d bytes for next "
              "offset\n",
              bytesLeft, (int)rdh.getOffsetNextPacket());
